requirements.txt
numpy==1.22.0
tensorflow==2.12.0
opencv-python==3.4.9.31
PyQt5==5.15.2
pyqt5-tools==5.15.2.3.1
face-recognition-models==0.3.0
Pillow==9.5.0
